Right, I really do apologise for the load of changes that seem to be coming along and may be really confusing. What i'll say, in the interest of trying to make this as simple as possible, Ignore the detail of this for now and concentrate on what you're actually trying to achieve with whatever you're writing, just be aware that this is here rather than jumping in and trying to understand it when you don't need to right now as you're not using it.
Milka pointed out to me earlier today that originally I had defined an abstract method getProperty(p, v) in GPSISDataMapper superclass  and that using this will most likely return multiple results. For example, PatientDMO.getProperty("first_name", "Milka"); would return everyone who has the first name 'Milka', so it's not a very useful method... 
Ideally we'd like to pass more than one parameter to the method so we can filter the results further :) So for example we can pass a Map where its Keys represent columns of the table (properties, fields whatever you want to call them) and the Value of that Key is the filter.
E.G:
A Map that looks like this:
	 * 	{	
	 * 		"first_name" => "VJ",
	 * 		"last_name" => "Patel"
	 * 	} 
will generate an SQL WHERE clause like this: WHERE first_name = 'VJ' AND last_name = 'Patel'

I've implemented this in my StaffMemberDMO, for now if you want to implement this for your DMO, copy what i've written in there and adjust the names and let me know if it struggles to work :p  Salman's going to look a bit more at Abstraction this weekend in the hope of actually abstracting these nicely so code doesn't get repeated (and you hopefully don't have to write this yourself!

Hava a good weekend :)